// src/app/models/sjcs/ScsDefendidosDesignasItem.ts
export class ScsDefendidosDesignasItem {
    idinstitucion: number;
    idturno: number;
    anio: number;
    numero: number;
    idpersona: string;
    fechamodificacion: Date;
    usumodificacion: number;
    observaciones: string;
    nombrerepresentante: string;
    calidad: string;
    idtipoencalidad: number;
    calidadinstitucion: number;
    constructor() { }
}